dImages/requirements.txt
Django==1.11.9
Pillow==5.0.0
psycopg2==2.7.3.2
django-imagekit==4.0.2
django-stdimage==3.0.0
tabulate==0.8.2
djangorestframework==3.7.7